+++ deployment/ConfigASDK.ps1
@@ -40,7 +40,8 @@
     * Supports usage in offline/disconnected environments
 
 .VERSION
-    1904.2  Support for App Service 1.6 (Coming soon)
+    1904.2  Support for App Service 1.6
+            Updated to support Azure Stack PowerShell 1.7.2 and AzureRM 2.5.0
     1904.1  Updated to support newer Ubuntu Server 16.04 image
             Updated with latest Windiws Updates and Adobe Flash Security Update downloads
             Check for Azure Stack Admin Login with correct privilege
@@ -1691,8 +1692,8 @@ try {
                 Get-PSRepository -Name "PSGallery"
                 Set-PSRepository -Name "PSGallery" -InstallationPolicy Trusted
                 Get-PSRepository -Name "PSGallery"
-                Install-Module -Name AzureStack -RequiredVersion 1.7.1 -Force -ErrorAction Stop
-                Install-Module AzureRM -RequiredVersion 2.4.0 -Force -ErrorAction Stop
+                Install-Module -Name AzureStack -RequiredVersion 1.7.2 -Force -ErrorAction Stop
+                Install-Module AzureRM -RequiredVersion 2.5.0 -Force -ErrorAction Stop
                 # Install the Azure.Storage module version 4.5.0
                 Install-Module -Name Azure.Storage -RequiredVersion 4.5.0 -Force -AllowClobber -Verbose
                 # Install the AzureRm.Storage module version 5.0.4