+++ changelog.md
@@ -6,7 +6,8 @@ Version History
 Here you'll find information on key changes, bug fixes and improvements made by version.
 
 ### What's new in version 1904.2?
-* Support for App Service 1.6 (Coming soon)
+* Support for App Service 1.6
+* Updated to support Azure Stack PowerShell 1.7.2 and AzureRM 2.5.0
 
 ### What's new in version 1904.1?
 * Updated to support newer Ubuntu Server 16.04 image